+++ js/components/modal.vue
@@ -8,7 +8,7 @@
                     <span aria-hidden="true">&times;</span>
                     <span class="sr-only" v-i18n="Close"></span>
                 </button>
-                <h4 class="modal-title" id="modal-title">{{title}}</h4>
+                <h4 class="modal-title" id="modal-title">{{{title}}}</h4>
             </div>
             <!-- Notifications -->
             <notification-zone class="modal-body"></notification-zone>