+++ js/front/organization/index.js
@@ -1,12 +1,9 @@
 /**
  * Organization display page JS module
  */
-// Catch all errors
-import 'front/bootstrap';
+import FrontMixin from 'front/mixin';
 
 import log from 'logger';
-import Auth from 'auth';
-import i18n from 'i18n';
 
 import Vue from 'vue';
 
@@ -26,7 +23,7 @@ Vue.options.replace = false;
 
 
 new Vue({
-    el: 'body',
+    mixins: [FrontMixin],
     components: {FollowButton, Tab, Tabset, ActivityTimeline, DashboardGraphs, SmallBox},
     data() {
         return {
@@ -40,15 +37,8 @@ new Vue({
         * Display the membership request modal
         */
         requestMembership(url) {
-            if (Auth.need_user(i18n._('You need to be logged in to request membership to an organization'))) {
-                return new Vue({
-                    mixins: [MembershipRequest],
-                    el: this.$els.modal,
-                    replace: false, // Needed while all components are not migrated to replace: true behavior
-                    parent: this,
-                    propsData: {url}
-                });
-            }
+            this.$auth(this._('You need to be logged in to request membership to an organization'));
+            this.$modal(MembershipRequest, {url});
         },
         showFollowers() {
             this.followersVisible = true;